Streamline routing; no longer care about final slash
Also redirect to expected URL if it does not match for some reason. This makes
it more obvious when the dev application is routing the production URL or vice
versa. In practice the redirect should never be triggered.

Finally use the WHATWG URL API instead of the deprecated `url.parse` function.