Resize network canvas to page; re-implement panning
The canvas that the network is drawn in changes size when the page is resized,
but also when the form to add a node is hidden or when alerts are shown/hidden.
This commit makes the canvas resize along with those changes, and ensures that
the network keeps being drawn to the center of the page.

In addition to that, panning is enabled in this commit. There turned out to not
be a conflict with the force-directed layout, but instead the D3 API is slightly
different than expected (cannot translate an element that the zoom/pan behavior
is called on, instead a separate container element must be translated).