Add NVvW demo
This is live at https://tcastermans.win.tue.nl/network at the moment. It is a
network game where participants can add their own node to a network that is seen
by all participants. An administrator can at any point in time percolate the
network, randomly deleting all edges with p = 0.5. This is then seen by
everyone. The game is used for the NVvW study day 2020, at 2020-11-07.

Detailed commits and messages can be found in the NVvW-demo branch. This commit
is a squashed merge of that branch at this point in time.