Indicate when nodes activate; fix possible hang
When positioning the nodes, the code could enter an infinite loop not being able
to find a location distant enough from other nodes. Decrease minimum distance
and put a maximum number of tries in place.

Bounce nodes when they are trying to activate, to visually indicate that.

Introduce some more constants for better maintainability. Decrease activation
rate and probability of a packet being spawned in a node.

Prevent `updateCanvas` being called twice when initializing the page.

Activate nodes at a (random) steady rate instead of after adding a packet,
deactivating or failing to activate.