Remove GET API (use socket.io only); tweaks
Only rely on socket.io for communication with the server. Get rid of the
deprecated API endpoints and code executing XMLHttpRequests except for the
`/getdata` endpoint.

Handle an incorrect secret by highlighting the input on the relevant room in the
overview instead of showing an alert.

Populate room overview using templating instead of through a separate request.