Contain socket.io updates to rooms
Also fix not being able to select nodes after a restart.

Updates about a network are now broadcast only to connections in the same room,
and also to connections without a room (index page). The latter is used to
update the number of nodes of a network in the rooms list. It can in the future
be used to update the rooms list (rooms added/deleted).