Get rid of `id` parameter; add a setting
This commit introduces a per-room setting to allow more than one node be added
per IP address. The reason for this change is two-fold: first, because of the
connection to the database, it was no longer feasible (properly) to have users
specify the ID externally. Secondly (most importantly!) the `id` parameter is
not user-friendly, the way it works is not transparent since node IDs are not
shown in the frontend. It would not be easy for a teacher to instruct their
students to use it. This new option is clear and easy to understand.